move CFEASolver files
